chore: rename "diagnoses" folder to "validators"
