+++ packages/info/lib/validators/config.js
