+++ packages/info/lib/validators/generic.js
